refactor(host-application): split dev server manager helpers
Extract AppService dev server lifecycle internals into focused orchestration, state, terminal, and process modules so changes stay easier to test and lower risk without changing the public behavior.
