--- apps/desktop/src-tauri/crates/host-application/src/app_service/dev_server_manager.rs
@@ -1,2172 +1,0 @@
-use super::{validate_hook_trust, AppService, DevServerEmitter, DevServerGroupRuntime};
-use anyhow::{anyhow, Context, Result};
-use host_domain::{
-    now_rfc3339, DevServerEvent, DevServerGroupState, DevServerScriptState, DevServerScriptStatus,
-    DevServerTerminalChunk,
-};
-use host_infra_system::{subprocess_path_env, RepoConfig, RepoDevServerScript};
-use std::collections::HashMap;
-#[cfg(unix)]
-use std::fs::File;
-use std::io::Read;
-#[cfg(unix)]
-use std::os::fd::FromRawFd;
-use std::process::{Command, Stdio};
-use std::sync::{Arc, Mutex};
-use std::thread;
-use std::time::Duration;
-
-const DEV_SERVER_TERMINAL_BUFFER_CHUNK_LIMIT: usize = 2_000;
-const DEV_SERVER_TERMINAL_BUFFER_BYTE_LIMIT: usize = 512 * 1024;
-const DEV_SERVER_TERMINAL_CHUNK_READ_SIZE: usize = 4 * 1024;
-const DEV_SERVER_STOP_TIMEOUT: Duration = Duration::from_secs(3);
-const DEV_SERVER_START_GRACE_PERIOD: Duration = Duration::from_millis(150);
-
-impl AppService {
-    pub fn dev_server_get_state(
-        &self,
-        repo_path: &str,
-        task_id: &str,
-    ) -> Result<DevServerGroupState> {
-        let repo_path = self.resolve_task_repo_path(repo_path)?;
-        let repo_config = self.config_store.repo_config(repo_path.as_str())?;
-        let worktree_path = self.resolve_dev_server_worktree_path(repo_path.as_str(), task_id);
-        let key = dev_server_group_key(repo_path.as_str(), task_id);
-        let mut groups = self
-            .dev_server_groups
-            .lock()
-            .map_err(|_| anyhow!("Dev server state lock poisoned"))?;
-        let runtime = groups.entry(key).or_insert_with(|| DevServerGroupRuntime {
-            state: build_group_state(
-                repo_path.as_str(),
-                task_id,
-                worktree_path.clone(),
-                &repo_config,
-            ),
-            emitter: None,
-        });
-        sync_group_state(
-            &mut runtime.state,
-            repo_path.as_str(),
-            task_id,
-            worktree_path,
-            &repo_config,
-        );
-        Ok(runtime.state.clone())
-    }
-
-    pub fn dev_server_start(
-        &self,
-        repo_path: &str,
-        task_id: &str,
-        emitter: DevServerEmitter,
-    ) -> Result<DevServerGroupState> {
-        let repo_path = self.resolve_task_repo_path(repo_path)?;
-        let repo_config = self.config_store.repo_config(repo_path.as_str())?;
-        if repo_config.dev_servers.is_empty() {
-            return Err(anyhow!(
-                "No builder dev server scripts are configured for {repo_path}. Add them in repository settings first."
-            ));
-        }
-        let worktree_path = self
-            .build_continuation_target_get(repo_path.as_str(), task_id)?
-            .ok_or_else(|| {
-                anyhow!(
-                    "Builder continuation cannot start until a builder worktree exists for task {task_id}. Start Builder first."
-                )
-            })?
-            .working_directory;
-        validate_hook_trust(repo_path.as_str(), &repo_config)?;
-        let key = dev_server_group_key(repo_path.as_str(), task_id);
-
-        {
-            let mut groups = self
-                .dev_server_groups
-                .lock()
-                .map_err(|_| anyhow!("Dev server state lock poisoned"))?;
-            let runtime = groups
-                .entry(key.clone())
-                .or_insert_with(|| DevServerGroupRuntime {
-                    state: build_group_state(
-                        repo_path.as_str(),
-                        task_id,
-                        Some(worktree_path.clone()),
-                        &repo_config,
-                    ),
-                    emitter: Some(emitter.clone()),
-                });
-            runtime.emitter = Some(emitter.clone());
-            sync_group_state(
-                &mut runtime.state,
-                repo_path.as_str(),
-                task_id,
-                Some(worktree_path.clone()),
-                &repo_config,
-            );
-            if runtime.state.scripts.iter().any(script_has_live_process) {
-                return Err(anyhow!(
-                    "Dev servers are already running for task {task_id}. Stop or restart them instead."
-                ));
-            }
-        }
-
-        emit_group_snapshot(self.dev_server_groups.clone(), &key);
-
-        let mut errors = Vec::new();
-        for script in repo_config.dev_servers.iter().cloned() {
-            match self.start_dev_server_script(
-                key.as_str(),
-                repo_path.as_str(),
-                task_id,
-                worktree_path.as_str(),
-                script,
-            ) {
-                Ok(_pid) => {}
-                Err(error) => errors.push(error.to_string()),
-            }
-        }
-
-        let state = self.dev_server_get_state(repo_path.as_str(), task_id)?;
-        emit_group_snapshot(self.dev_server_groups.clone(), &key);
-        if errors.is_empty() || state.scripts.iter().any(script_has_live_process) {
-            Ok(state)
-        } else {
-            Err(anyhow!(errors.join("\n")))
-        }
-    }
-
-    pub fn dev_server_stop(&self, repo_path: &str, task_id: &str) -> Result<DevServerGroupState> {
-        let repo_path = self.resolve_task_repo_path(repo_path)?;
-        let repo_config = self.config_store.repo_config(repo_path.as_str())?;
-        let key = dev_server_group_key(repo_path.as_str(), task_id);
-        {
-            let mut groups = self
-                .dev_server_groups
-                .lock()
-                .map_err(|_| anyhow!("Dev server state lock poisoned"))?;
-            let runtime = groups
-                .entry(key.clone())
-                .or_insert_with(|| DevServerGroupRuntime {
-                    state: build_group_state(repo_path.as_str(), task_id, None, &repo_config),
-                    emitter: None,
-                });
-            sync_group_state(
-                &mut runtime.state,
-                repo_path.as_str(),
-                task_id,
-                None,
-                &repo_config,
-            );
-        }
-
-        let targets = self.mark_dev_servers_stopping(key.as_str())?;
-        if targets.is_empty() {
-            let state = self.dev_server_get_state(repo_path.as_str(), task_id)?;
-            emit_group_snapshot(self.dev_server_groups.clone(), &key);
-            return Ok(state);
-        }
-
-        let mut errors = Vec::new();
-        for (script_id, pid) in targets {
-            if let Err(error) = stop_process_group(pid, DEV_SERVER_STOP_TIMEOUT) {
-                self.mark_dev_server_stop_failed(key.as_str(), &script_id, pid, &error.to_string());
-                errors.push(format!("Failed stopping dev server {script_id}: {error}"));
-                continue;
-            }
-            self.mark_dev_server_stopped(key.as_str(), &script_id, pid, None);
-        }
-
-        let state = self.dev_server_get_state(repo_path.as_str(), task_id)?;
-        emit_group_snapshot(self.dev_server_groups.clone(), &key);
-        if errors.is_empty() {
-            Ok(state)
-        } else {
-            Err(anyhow!(errors.join("\n")))
-        }
-    }
-
-    pub fn dev_server_restart(
-        &self,
-        repo_path: &str,
-        task_id: &str,
-        emitter: DevServerEmitter,
-    ) -> Result<DevServerGroupState> {
-        self.dev_server_stop(repo_path, task_id)?;
-        self.dev_server_start(repo_path, task_id, emitter)
-    }
-
-    pub(crate) fn stop_dev_servers_for_task(&self, repo_path: &str, task_id: &str) -> Result<()> {
-        let _ = self.dev_server_stop(repo_path, task_id)?;
-        Ok(())
-    }
-
-    pub(crate) fn stop_all_dev_servers(&self) -> Result<()> {
-        let targets = self
-            .dev_server_groups
-            .lock()
-            .map_err(|_| anyhow!("Dev server state lock poisoned"))?
-            .values()
-            .map(|group| (group.state.repo_path.clone(), group.state.task_id.clone()))
-            .collect::<Vec<_>>();
-        let mut errors = Vec::new();
-        for (repo_path, task_id) in targets {
-            if let Err(error) = self.stop_dev_servers_for_task(repo_path.as_str(), task_id.as_str())
-            {
-                errors.push(format!(
-                    "Failed stopping dev servers for {}:{}: {error:#}",
-                    repo_path, task_id
-                ));
-            }
-        }
-
-        if errors.is_empty() {
-            Ok(())
-        } else {
-            Err(anyhow!(errors.join("\n")))
-        }
-    }
-
-    fn start_dev_server_script(
-        &self,
-        group_key: &str,
-        repo_path: &str,
-        task_id: &str,
-        worktree_path: &str,
-        script: RepoDevServerScript,
-    ) -> Result<u32> {
-        let script_display_name = sanitize_dev_server_display_name(script.name.as_str());
-        self.update_script_state(group_key, script.id.as_str(), |state| {
-            state.status = DevServerScriptStatus::Starting;
-            state.pid = None;
-            state.started_at = None;
-            state.exit_code = None;
-            state.last_error = None;
-            state.buffered_terminal_chunks.clear();
-        })?;
-        self.append_terminal_system_message(
-            group_key,
-            repo_path,
-            task_id,
-            script.id.as_str(),
-            format!("Starting `{}`", script.command),
-        );
-
-        let mut spawned_pid = None;
-        let mut start_failure_already_recorded = false;
-        let start_result = (|| -> Result<u32> {
-            let mut command =
-                build_dev_server_command(script.command.as_str(), script.name.as_str())?;
-            command.current_dir(worktree_path);
-            let path_value = subprocess_path_env().ok_or_else(|| {
-                anyhow!(
-                    "Failed to assemble subprocess PATH for dev server `{}`",
-                    script.name
-                )
-            })?;
-            command.env("PATH", path_value);
-            configure_process_group(&mut command);
-
-            #[cfg(unix)]
-            let terminal_reader: Box<dyn Read + Send> = {
-                let (terminal_reader, terminal_writer) = open_pty_pair()?;
-                let stdout_writer = terminal_writer
-                    .try_clone()
-                    .context("Failed duplicating pseudo terminal handle for dev server stdout")?;
-                command.stdout(Stdio::from(stdout_writer));
-                command.stderr(Stdio::from(terminal_writer));
-                Box::new(terminal_reader)
-            };
-
-            #[cfg(not(unix))]
-            return Err(anyhow!(
-                "Builder dev servers require PTY-backed terminal capture and are only supported on Unix hosts in this build."
-            ));
-
-            #[cfg(unix)]
-            let mut child = command.spawn().with_context(|| {
-                format!(
-                    "Failed to start dev server {} in {} using command `{}`",
-                    script_display_name, worktree_path, script.command
-                )
-            })?;
-
-            #[cfg(not(unix))]
-            let mut child = command.spawn().with_context(|| {
-                format!(
-                    "Failed to start dev server {} in {} using command `{}`",
-                    script_display_name, worktree_path, script.command
-                )
-            })?;
-            let pid = child.id();
-            spawned_pid = Some(pid);
-
-            #[cfg(unix)]
-            if let Err(error) = spawn_dev_server_parent_death_watcher(std::process::id(), pid) {
-                eprintln!(
-                    "OpenDucktor warning: failed to attach dev server parent-death watcher for pid {}: {error:#}",
-                    pid
-                );
-            }
-
-            let started_at = now_rfc3339();
-
-            spawn_terminal_forwarder(
-                self.dev_server_groups.clone(),
-                group_key.to_string(),
-                repo_path.to_string(),
-                task_id.to_string(),
-                script.id.clone(),
-                terminal_reader,
-            );
-
-            self.update_script_state(group_key, script.id.as_str(), |state| {
-                state.status = DevServerScriptStatus::Starting;
-                state.pid = Some(pid);
-                state.started_at = Some(started_at.clone());
-                state.exit_code = None;
-                state.last_error = None;
-            })?;
-
-            if let Some(status) =
-                wait_for_immediate_dev_server_exit(&mut child, DEV_SERVER_START_GRACE_PERIOD)?
-            {
-                let message = dev_server_exit_message(status.code());
-                self.mark_dev_server_start_exit_failed(
-                    group_key,
-                    repo_path,
-                    task_id,
-                    script.id.as_str(),
-                    status.code(),
-                    message.as_str(),
-                );
-                spawned_pid = None;
-                start_failure_already_recorded = true;
-                return Err(anyhow!(message));
-            }
-
-            self.update_script_state(group_key, script.id.as_str(), |state| {
-                state.status = DevServerScriptStatus::Running;
-                state.pid = Some(pid);
-                state.started_at = Some(started_at.clone());
-                state.exit_code = None;
-                state.last_error = None;
-            })?;
-
-            spawn_waiter(
-                self.dev_server_groups.clone(),
-                group_key.to_string(),
-                repo_path.to_string(),
-                task_id.to_string(),
-                script.id.clone(),
-                pid,
-                child,
-            );
-            Ok(pid)
-        })();
-
-        if let Err(error) = start_result {
-            let mut message = format!("{error:#}");
-            if let Some(pid) = spawned_pid {
-                if let Err(stop_error) = stop_process_group(pid, DEV_SERVER_STOP_TIMEOUT) {
-                    message = format!(
-                        "{message}\nFailed stopping partially started dev server {} (pid {pid}): {stop_error:#}",
-                        script_display_name
-                    );
-                }
-            }
-            if !start_failure_already_recorded {
-                self.mark_dev_server_start_failed(
-                    group_key,
-                    repo_path,
-                    task_id,
-                    script.id.as_str(),
-                    message.as_str(),
-                );
-            }
-            return Err(anyhow!(message));
-        }
-
-        start_result
-    }
-
-    fn mark_dev_servers_stopping(&self, group_key: &str) -> Result<Vec<(String, u32)>> {
-        let mut groups = self
-            .dev_server_groups
-            .lock()
-            .map_err(|_| anyhow!("Dev server state lock poisoned"))?;
-        let runtime = groups
-            .get_mut(group_key)
-            .ok_or_else(|| anyhow!("Dev server state missing for active task"))?;
-        let updated_at = now_rfc3339();
-        runtime.state.updated_at = updated_at.clone();
-        let mut targets = Vec::new();
-        let mut changed_scripts = Vec::new();
-        for script in &mut runtime.state.scripts {
-            script.buffered_terminal_chunks.clear();
-            if let Some(pid) = script.pid {
-                script.status = DevServerScriptStatus::Stopping;
-                script.last_error = None;
-                targets.push((script.script_id.clone(), pid));
-                changed_scripts.push(script.clone());
-            } else if script.status != DevServerScriptStatus::Stopped
-                || script.last_error.is_some()
-                || script.exit_code.is_some()
-            {
-                script.status = DevServerScriptStatus::Stopped;
-                script.started_at = None;
-                script.exit_code = None;
-                script.last_error = None;
-                changed_scripts.push(script.clone());
-            }
-        }
-        let emitter = runtime.emitter.clone();
-        let repo_path = runtime.state.repo_path.clone();
-        let task_id = runtime.state.task_id.clone();
-        drop(groups);
-        if let Some(emitter) = emitter {
-            for script in changed_scripts {
-                emitter(DevServerEvent::ScriptStatusChanged {
-                    repo_path: repo_path.clone(),
-                    task_id: task_id.clone(),
-                    script,
-                    updated_at: updated_at.clone(),
-                });
-            }
-        }
-        Ok(targets)
-    }
-
-    fn mark_dev_server_stopped(
-        &self,
-        group_key: &str,
-        script_id: &str,
-        pid: u32,
-        exit_code: Option<i32>,
-    ) {
-        let _ = self.update_script_state(group_key, script_id, |state| {
-            if state.pid != Some(pid) {
-                return;
-            }
-            state.status = DevServerScriptStatus::Stopped;
-            state.pid = None;
-            state.started_at = None;
-            state.exit_code = exit_code;
-            state.last_error = None;
-        });
-    }
-
-    fn mark_dev_server_stop_failed(
-        &self,
-        group_key: &str,
-        script_id: &str,
-        pid: u32,
-        message: &str,
-    ) {
-        let _ = self.update_script_state(group_key, script_id, |state| {
-            if state.pid != Some(pid) {
-                return;
-            }
-            state.status = DevServerScriptStatus::Failed;
-            state.last_error = Some(message.to_string());
-        });
-    }
-
-    fn mark_dev_server_start_failed(
-        &self,
-        group_key: &str,
-        repo_path: &str,
-        task_id: &str,
-        script_id: &str,
-        message: &str,
-    ) {
-        let _ = self.update_script_state(group_key, script_id, |state| {
-            state.status = DevServerScriptStatus::Failed;
-            state.pid = None;
-            state.started_at = None;
-            state.exit_code = None;
-            state.last_error = Some(message.to_string());
-        });
-        self.append_terminal_system_message(group_key, repo_path, task_id, script_id, message);
-    }
-
-    fn mark_dev_server_start_exit_failed(
-        &self,
-        group_key: &str,
-        repo_path: &str,
-        task_id: &str,
-        script_id: &str,
-        exit_code: Option<i32>,
-        message: &str,
-    ) {
-        let _ = self.update_script_state(group_key, script_id, |state| {
-            state.status = DevServerScriptStatus::Failed;
-            state.pid = None;
-            state.started_at = None;
-            state.exit_code = exit_code;
-            state.last_error = Some(message.to_string());
-        });
-        self.append_terminal_system_message(group_key, repo_path, task_id, script_id, message);
-    }
-
-    fn update_script_state<F>(&self, group_key: &str, script_id: &str, update: F) -> Result<()>
-    where
-        F: FnOnce(&mut DevServerScriptState),
-    {
-        let mut groups = self
-            .dev_server_groups
-            .lock()
-            .map_err(|_| anyhow!("Dev server state lock poisoned"))?;
-        let runtime = groups
-            .get_mut(group_key)
-            .ok_or_else(|| anyhow!("Dev server state missing for active task"))?;
-        let script_snapshot = {
-            let script = runtime
-                .state
-                .scripts
-                .iter_mut()
-                .find(|script| script.script_id == script_id)
-                .ok_or_else(|| anyhow!("Unknown dev server script: {script_id}"))?;
-            update(script);
-            script.clone()
-        };
-        runtime.state.updated_at = now_rfc3339();
-        let updated_at = runtime.state.updated_at.clone();
-        let emitter = runtime.emitter.clone();
-        let repo_path = runtime.state.repo_path.clone();
-        let task_id = runtime.state.task_id.clone();
-        drop(groups);
-        if let Some(emitter) = emitter {
-            emitter(DevServerEvent::ScriptStatusChanged {
-                repo_path,
-                task_id,
-                script: script_snapshot,
-                updated_at,
-            });
-        }
-        Ok(())
-    }
-
-    fn resolve_dev_server_worktree_path(&self, repo_path: &str, task_id: &str) -> Option<String> {
-        self.build_continuation_target_get(repo_path, task_id)
-            .ok()
-            .and_then(|target| target.map(|entry| entry.working_directory))
-    }
-
-    fn append_terminal_system_message(
-        &self,
-        group_key: &str,
-        repo_path: &str,
-        task_id: &str,
-        script_id: &str,
-        message: impl AsRef<str>,
-    ) {
-        emit_terminal_chunk(
-            &self.dev_server_groups,
-            group_key,
-            repo_path,
-            task_id,
-            script_id,
-            format_terminal_system_message(message.as_ref()),
-        );
-    }
-}
-
-fn build_group_state(
-    repo_path: &str,
-    task_id: &str,
-    worktree_path: Option<String>,
-    repo_config: &RepoConfig,
-) -> DevServerGroupState {
-    DevServerGroupState {
-        repo_path: repo_path.to_string(),
-        task_id: task_id.to_string(),
-        worktree_path,
-        scripts: repo_config
-            .dev_servers
-            .iter()
-            .map(new_script_state)
-            .collect(),
-        updated_at: now_rfc3339(),
-    }
-}
-
-fn sync_group_state(
-    state: &mut DevServerGroupState,
-    repo_path: &str,
-    task_id: &str,
-    worktree_path: Option<String>,
-    repo_config: &RepoConfig,
-) {
-    state.repo_path = repo_path.to_string();
-    state.task_id = task_id.to_string();
-    state.worktree_path = worktree_path;
-
-    let mut existing = state
-        .scripts
-        .drain(..)
-        .map(|script| (script.script_id.clone(), script))
-        .collect::<HashMap<_, _>>();
-    let mut next_scripts = Vec::with_capacity(repo_config.dev_servers.len());
-    for repo_script in &repo_config.dev_servers {
-        if let Some(mut script) = existing.remove(repo_script.id.as_str()) {
-            script.name = repo_script.name.clone();
-            script.command = repo_script.command.clone();
-            next_scripts.push(script);
-        } else {
-            next_scripts.push(new_script_state(repo_script));
-        }
-    }
-    next_scripts.extend(existing.into_values().filter(script_has_live_process));
-    state.scripts = next_scripts;
-    state.updated_at = now_rfc3339();
-}
-
-fn new_script_state(script: &RepoDevServerScript) -> DevServerScriptState {
-    DevServerScriptState {
-        script_id: script.id.clone(),
-        name: script.name.clone(),
-        command: script.command.clone(),
-        status: DevServerScriptStatus::Stopped,
-        pid: None,
-        started_at: None,
-        exit_code: None,
-        last_error: None,
-        buffered_terminal_chunks: Vec::new(),
-        next_terminal_sequence: 0,
-    }
-}
-
-fn script_has_live_process(script: &DevServerScriptState) -> bool {
-    script.pid.is_some()
-}
-
-fn dev_server_group_key(repo_path: &str, task_id: &str) -> String {
-    format!("{repo_path}::{task_id}")
-}
-
-fn emit_group_snapshot(
-    groups: Arc<Mutex<HashMap<String, DevServerGroupRuntime>>>,
-    group_key: &str,
-) {
-    let (emitter, state) = {
-        let Ok(groups) = groups.lock() else {
-            return;
-        };
-        let Some(runtime) = groups.get(group_key) else {
-            return;
-        };
-        (runtime.emitter.clone(), runtime.state.clone())
-    };
-    if let Some(emitter) = emitter {
-        emitter(DevServerEvent::Snapshot { state });
-    }
-}
-
-fn emit_terminal_chunk(
-    groups: &Arc<Mutex<HashMap<String, DevServerGroupRuntime>>>,
-    group_key: &str,
-    repo_path: &str,
-    task_id: &str,
-    script_id: &str,
-    data: String,
-) {
-    if data.is_empty() {
-        return;
-    }
-
-    let timestamp = now_rfc3339();
-    let (emitter, terminal_chunk) = {
-        let Ok(mut groups) = groups.lock() else {
-            return;
-        };
-        let Some(runtime) = groups.get_mut(group_key) else {
-            return;
-        };
-        let Some(script) = runtime
-            .state
-            .scripts
-            .iter_mut()
-            .find(|script| script.script_id == script_id)
-        else {
-            return;
-        };
-        let terminal_chunk = push_terminal_chunk(script, data, timestamp);
-        runtime.state.updated_at = now_rfc3339();
-        (runtime.emitter.clone(), terminal_chunk)
-    };
-
-    if let Some(emitter) = emitter {
-        emitter(DevServerEvent::TerminalChunk {
-            repo_path: repo_path.to_string(),
-            task_id: task_id.to_string(),
-            terminal_chunk,
-        });
-    }
-}
-
-fn spawn_terminal_forwarder<R>(
-    groups: Arc<Mutex<HashMap<String, DevServerGroupRuntime>>>,
-    group_key: String,
-    repo_path: String,
-    task_id: String,
-    script_id: String,
-    reader: R,
-) where
-    R: Read + Send + 'static,
-{
-    std::thread::spawn(move || {
-        let mut reader = reader;
-        let mut buffer = [0_u8; DEV_SERVER_TERMINAL_CHUNK_READ_SIZE];
-        let mut pending = Vec::new();
-
-        loop {
-            match reader.read(&mut buffer) {
-                Ok(0) => {
-                    let _ = flush_pending_terminal_bytes(
-                        &groups,
-                        group_key.as_str(),
-                        repo_path.as_str(),
-                        task_id.as_str(),
-                        script_id.as_str(),
-                        &mut pending,
-                        true,
-                    );
-                    break;
-                }
-                Ok(read) => {
-                    pending.extend_from_slice(&buffer[..read]);
-                    if !flush_pending_terminal_bytes(
-                        &groups,
-                        group_key.as_str(),
-                        repo_path.as_str(),
-                        task_id.as_str(),
-                        script_id.as_str(),
-                        &mut pending,
-                        false,
-                    ) {
-                        break;
-                    }
-                }
-                Err(error) => {
-                    if error.kind() == std::io::ErrorKind::Interrupted {
-                        continue;
-                    }
-                    emit_terminal_chunk(
-                        &groups,
-                        group_key.as_str(),
-                        repo_path.as_str(),
-                        task_id.as_str(),
-                        script_id.as_str(),
-                        format_terminal_system_message(
-                            format!("Dev server terminal stream failed: {error}").as_str(),
-                        ),
-                    );
-                    break;
-                }
-            }
-        }
-    });
-}
-
-fn flush_pending_terminal_bytes(
-    groups: &Arc<Mutex<HashMap<String, DevServerGroupRuntime>>>,
-    group_key: &str,
-    repo_path: &str,
-    task_id: &str,
-    script_id: &str,
-    pending: &mut Vec<u8>,
-    reached_eof: bool,
-) -> bool {
-    loop {
-        if pending.is_empty() {
-            return true;
-        }
-
-        match std::str::from_utf8(pending) {
-            Ok(text) => {
-                emit_terminal_chunk(
-                    groups,
-                    group_key,
-                    repo_path,
-                    task_id,
-                    script_id,
-                    text.to_string(),
-                );
-                pending.clear();
-                return true;
-            }
-            Err(error) => {
-                let valid_up_to = error.valid_up_to();
-                if valid_up_to > 0 {
-                    let text = std::str::from_utf8(&pending[..valid_up_to])
-                        .expect("valid UTF-8 prefix should decode");
-                    emit_terminal_chunk(
-                        groups,
-                        group_key,
-                        repo_path,
-                        task_id,
-                        script_id,
-                        text.to_string(),
-                    );
-                    pending.drain(0..valid_up_to);
-                    continue;
-                }
-
-                if error.error_len().is_none() && !reached_eof {
-                    return true;
-                }
-
-                if let Some(error_len) = error.error_len() {
-                    emit_terminal_chunk(
-                        groups,
-                        group_key,
-                        repo_path,
-                        task_id,
-                        script_id,
-                        format_terminal_system_message(
-                            "Dev server terminal output contained invalid UTF-8 bytes and could not be fully rendered.",
-                        ),
-                    );
-                    pending.drain(0..error_len);
-                    continue;
-                }
-
-                emit_terminal_chunk(
-                    groups,
-                    group_key,
-                    repo_path,
-                    task_id,
-                    script_id,
-                    format_terminal_system_message(
-                        "Dev server terminal output contained invalid UTF-8 bytes and could not be fully rendered.",
-                    ),
-                );
-                pending.clear();
-                return true;
-            }
-        }
-    }
-}
-
-fn push_terminal_chunk(
-    script: &mut DevServerScriptState,
-    data: String,
-    timestamp: String,
-) -> DevServerTerminalChunk {
-    let terminal_chunk = DevServerTerminalChunk {
-        script_id: script.script_id.clone(),
-        sequence: script.next_terminal_sequence,
-        data,
-        timestamp,
-    };
-    script.next_terminal_sequence = script.next_terminal_sequence.saturating_add(1);
-    script.buffered_terminal_chunks.push(terminal_chunk.clone());
-    trim_terminal_chunk_buffer(&mut script.buffered_terminal_chunks);
-    terminal_chunk
-}
-
-fn trim_terminal_chunk_buffer(chunks: &mut Vec<DevServerTerminalChunk>) {
-    let mut remove_count = 0;
-    let mut total_bytes = chunks.iter().map(|chunk| chunk.data.len()).sum::<usize>();
-
-    while chunks.len().saturating_sub(remove_count) > DEV_SERVER_TERMINAL_BUFFER_CHUNK_LIMIT
-        || total_bytes > DEV_SERVER_TERMINAL_BUFFER_BYTE_LIMIT
-    {
-        let Some(chunk) = chunks.get(remove_count) else {
-            break;
-        };
-        total_bytes = total_bytes.saturating_sub(chunk.data.len());
-        remove_count += 1;
-    }
-
-    if remove_count > 0 {
-        chunks.drain(0..remove_count);
-    }
-}
-
-fn format_terminal_system_message(message: &str) -> String {
-    let normalized = message.replace("\r\n", "\n").replace('\n', "\r\n");
-    if normalized.ends_with("\r\n") {
-        normalized
-    } else {
-        format!("{normalized}\r\n")
-    }
-}
-
-fn wait_for_immediate_dev_server_exit(
-    child: &mut std::process::Child,
-    timeout: Duration,
-) -> Result<Option<std::process::ExitStatus>> {
-    let poll_interval = Duration::from_millis(25);
-    let mut waited = Duration::ZERO;
-    loop {
-        if let Some(status) = child.try_wait()? {
-            return Ok(Some(status));
-        }
-        if waited >= timeout {
-            return Ok(None);
-        }
-        thread::sleep(poll_interval);
-        waited = waited.saturating_add(poll_interval);
-    }
-}
-
-fn dev_server_exit_message(exit_code: Option<i32>) -> String {
-    if let Some(code) = exit_code {
-        format!("Dev server exited with code {code}.")
-    } else {
-        "Dev server exited after receiving a signal.".to_string()
-    }
-}
-
-fn sanitize_dev_server_display_name(script_name: &str) -> String {
-    const DISPLAY_NAME_LIMIT: usize = 80;
-
-    let mut sanitized = String::new();
-    for ch in script_name.chars() {
-        match ch {
-            '\n' => sanitized.push_str("\\n"),
-            '\r' => sanitized.push_str("\\r"),
-            '\t' => sanitized.push_str("\\t"),
-            _ if ch.is_control() => sanitized.push(' '),
-            _ => sanitized.push(ch),
-        }
-
-        if sanitized.chars().count() > DISPLAY_NAME_LIMIT {
-            sanitized = sanitized.chars().take(DISPLAY_NAME_LIMIT).collect();
-            sanitized.push_str("...");
-            return sanitized;
-        }
-    }
-
-    sanitized.trim().to_string()
-}
-
-#[cfg(unix)]
-fn build_dev_server_command(command: &str, script_name: &str) -> Result<Command> {
-    if command.trim().is_empty() {
-        let script_name = sanitize_dev_server_display_name(script_name);
-        return Err(anyhow!(
-            "Dev server command is empty for {}. Provide a command to run.",
-            script_name
-        ));
-    }
-
-    let mut process = Command::new("/bin/sh");
-    process.arg("-c");
-    process.arg(command);
-    Ok(process)
-}
-
-#[cfg(not(unix))]
-fn build_dev_server_command(command: &str, script_name: &str) -> Result<Command> {
-    let script_name = sanitize_dev_server_display_name(script_name);
-    let parsed = shell_words::split(command).map_err(|error| {
-        anyhow!(
-            "Invalid dev server command syntax for {}. Use argv tokens, or explicitly invoke a shell (for example: sh -lc '...'): {error}",
-            script_name
-        )
-    })?;
-    let (program, args) = parsed.split_first().ok_or_else(|| {
-        anyhow!(
-            "Dev server command is empty for {}. Provide an executable name.",
-            script_name
-        )
-    })?;
-
-    let mut process = Command::new(program);
-    process.args(args);
-    Ok(process)
-}
-
-#[cfg(unix)]
-fn open_pty_pair() -> Result<(File, File)> {
-    let mut master_fd = 0;
-    let mut slave_fd = 0;
-    let result = unsafe {
-        libc::openpty(
-            &mut master_fd,
-            &mut slave_fd,
-            std::ptr::null_mut(),
-            std::ptr::null_mut(),
-            std::ptr::null_mut(),
-        )
-    };
-    if result != 0 {
-        return Err(anyhow!(
-            "Failed creating pseudo terminal for dev server logs: {}",
-            std::io::Error::last_os_error()
-        ));
-    }
-
-    let master = unsafe { File::from_raw_fd(master_fd) };
-    let slave = unsafe { File::from_raw_fd(slave_fd) };
-    Ok((master, slave))
-}
-
-fn spawn_waiter(
-    groups: Arc<Mutex<HashMap<String, DevServerGroupRuntime>>>,
-    group_key: String,
-    repo_path: String,
-    task_id: String,
-    script_id: String,
-    pid: u32,
-    mut child: std::process::Child,
-) {
-    std::thread::spawn(move || {
-        let status = child.wait();
-        let (exit_code, message) = match status {
-            Ok(status) => (status.code(), dev_server_exit_message(status.code())),
-            Err(error) => (None, format!("Failed waiting for dev server exit: {error}")),
-        };
-
-        let mut emitted_terminal_chunk = None;
-        let (emitter, script, updated_at) = {
-            let Ok(mut groups) = groups.lock() else {
-                return;
-            };
-            let Some(runtime) = groups.get_mut(group_key.as_str()) else {
-                return;
-            };
-            let Some(script) = runtime
-                .state
-                .scripts
-                .iter_mut()
-                .find(|script| script.script_id == script_id)
-            else {
-                return;
-            };
-            if script.pid != Some(pid) {
-                return;
-            }
-            let expected_stop = matches!(script.status, DevServerScriptStatus::Stopping);
-            script.pid = None;
-            script.started_at = None;
-            script.exit_code = exit_code;
-            if expected_stop {
-                script.status = DevServerScriptStatus::Stopped;
-                script.last_error = None;
-            } else {
-                script.status = DevServerScriptStatus::Failed;
-                script.last_error = Some(message.clone());
-                emitted_terminal_chunk = Some(push_terminal_chunk(
-                    script,
-                    format_terminal_system_message(message.as_str()),
-                    now_rfc3339(),
-                ));
-            }
-            runtime.state.updated_at = now_rfc3339();
-            (
-                runtime.emitter.clone(),
-                script.clone(),
-                runtime.state.updated_at.clone(),
-            )
-        };
-
-        if let Some(emitter) = emitter {
-            if let Some(terminal_chunk) = emitted_terminal_chunk {
-                emitter(DevServerEvent::TerminalChunk {
-                    repo_path: repo_path.clone(),
-                    task_id: task_id.clone(),
-                    terminal_chunk,
-                });
-            }
-            emitter(DevServerEvent::ScriptStatusChanged {
-                repo_path,
-                task_id,
-                script,
-                updated_at,
-            });
-        }
-    });
-}
-
-#[cfg(unix)]
-fn configure_process_group(command: &mut Command) {
-    use std::os::unix::process::CommandExt;
-    command.process_group(0);
-}
-
-#[cfg(not(unix))]
-fn configure_process_group(_command: &mut Command) {}
-
-#[cfg(unix)]
-fn spawn_dev_server_parent_death_watcher(parent_pid: u32, child_pid: u32) -> Result<()> {
-    let watcher_script = format!(
-        r#"P={parent_pid}; C={child_pid}; while kill -0 "$P" 2>/dev/null && kill -0 "$C" 2>/dev/null; do sleep 1; done; if ! kill -0 "$P" 2>/dev/null && kill -0 "$C" 2>/dev/null; then kill -TERM -"$C" 2>/dev/null || true; sleep 1; kill -KILL -"$C" 2>/dev/null || true; kill -KILL "$C" 2>/dev/null || true; fi"#
-    );
-    Command::new("/bin/sh")
-        .arg("-c")
-        .arg(watcher_script)
-        .stdin(Stdio::null())
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .spawn()
-        .context("Failed to spawn dev server parent-death watcher")?;
-    Ok(())
-}
-
-#[cfg(not(unix))]
-fn spawn_dev_server_parent_death_watcher(_parent_pid: u32, _child_pid: u32) -> Result<()> {
-    Ok(())
-}
-
-#[cfg(unix)]
-fn stop_process_group(pid: u32, timeout: Duration) -> Result<()> {
-    let pid =
-        i32::try_from(pid).map_err(|_| anyhow!("Invalid process id for dev server: {pid}"))?;
-    if pid <= 0 {
-        return Err(anyhow!("Invalid process id for dev server: {pid}"));
-    }
-    let wait_pid = pid as u32;
-    let term_result = unsafe { libc::killpg(pid, libc::SIGTERM) };
-    if term_result != 0 {
-        let error = std::io::Error::last_os_error();
-        if error.raw_os_error() == Some(libc::ESRCH) {
-            return Ok(());
-        }
-        return Err(anyhow!(
-            "Failed sending SIGTERM to process group {pid}: {error}"
-        ));
-    }
-    if super::wait_for_process_exit_by_pid(wait_pid, timeout) {
-        return Ok(());
-    }
-    let kill_result = unsafe { libc::killpg(pid, libc::SIGKILL) };
-    if kill_result != 0 {
-        let error = std::io::Error::last_os_error();
-        if error.raw_os_error() == Some(libc::ESRCH) {
-            return Ok(());
-        }
-        return Err(anyhow!(
-            "Failed sending SIGKILL to process group {pid}: {error}"
-        ));
-    }
-    if super::wait_for_process_exit_by_pid(wait_pid, timeout) {
-        return Ok(());
-    }
-    Err(anyhow!(
-        "Timed out waiting for process group {pid} to stop after SIGTERM and SIGKILL"
-    ))
-}
-
-#[cfg(not(unix))]
-fn stop_process_group(pid: u32, timeout: Duration) -> Result<()> {
-    let _ = (pid, timeout);
-    Err(anyhow!(
-        "Builder dev servers are only supported on Unix hosts in this build."
-    ))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::app_service::test_support::{
-        build_service_with_state, build_service_with_store, init_git_repo, lock_env, make_task,
-        set_env_var, spawn_sleep_process_group, unique_temp_path, wait_for_path_exists,
-        wait_for_process_exit, write_executable_script,
-    };
-    use crate::app_service::{HookTrustConfirmationPort, HookTrustConfirmationRequest};
-    use host_domain::{GitCurrentBranch, TaskStatus};
-    use host_infra_system::AppConfigStore;
-    use std::collections::VecDeque;
-    use std::fs;
-    use std::io::{self, Read};
-    use std::sync::{Arc, Mutex};
-    use std::thread;
-    use std::time::Duration;
-
-    struct AllowHookTrustConfirmationPort;
-
-    impl HookTrustConfirmationPort for AllowHookTrustConfirmationPort {
-        fn confirm_trusted_hooks(&self, _request: &HookTrustConfirmationRequest) -> Result<()> {
-            Ok(())
-        }
-    }
-
-    fn repo_config(dev_servers: Vec<RepoDevServerScript>) -> RepoConfig {
-        RepoConfig {
-            dev_servers,
-            ..Default::default()
-        }
-    }
-
-    enum ScriptedReadStep {
-        Chunk(Vec<u8>),
-        Error(io::ErrorKind),
-        Eof,
-    }
-
-    struct ScriptedReader {
-        steps: VecDeque<ScriptedReadStep>,
-    }
-
-    impl ScriptedReader {
-        fn new(steps: Vec<ScriptedReadStep>) -> Self {
-            Self {
-                steps: steps.into(),
-            }
-        }
-    }
-
-    impl Read for ScriptedReader {
-        fn read(&mut self, buffer: &mut [u8]) -> io::Result<usize> {
-            match self.steps.pop_front().unwrap_or(ScriptedReadStep::Eof) {
-                ScriptedReadStep::Chunk(bytes) => {
-                    let read = bytes.len().min(buffer.len());
-                    buffer[..read].copy_from_slice(&bytes[..read]);
-                    Ok(read)
-                }
-                ScriptedReadStep::Error(kind) => Err(io::Error::from(kind)),
-                ScriptedReadStep::Eof => Ok(0),
-            }
-        }
-    }
-
-    fn build_runtime_groups_for_forwarder_test(
-    ) -> Arc<Mutex<HashMap<String, DevServerGroupRuntime>>> {
-        Arc::new(Mutex::new(HashMap::from([(
-            "repo::task-forwarder".to_string(),
-            DevServerGroupRuntime {
-                state: DevServerGroupState {
-                    repo_path: "repo".to_string(),
-                    task_id: "task-forwarder".to_string(),
-                    worktree_path: Some("/tmp/worktree".to_string()),
-                    scripts: vec![DevServerScriptState {
-                        script_id: "server-1".to_string(),
-                        name: "Server".to_string(),
-                        command: "bun run dev".to_string(),
-                        status: DevServerScriptStatus::Running,
-                        pid: Some(99),
-                        started_at: Some("2026-03-19T10:00:00Z".to_string()),
-                        exit_code: None,
-                        last_error: None,
-                        buffered_terminal_chunks: Vec::new(),
-                        next_terminal_sequence: 0,
-                    }],
-                    updated_at: "2026-03-19T10:00:00Z".to_string(),
-                },
-                emitter: None,
-            },
-        )])))
-    }
-
-    fn read_forwarder_output(
-        groups: &Arc<Mutex<HashMap<String, DevServerGroupRuntime>>>,
-    ) -> String {
-        groups
-            .lock()
-            .expect("group lock poisoned")
-            .get("repo::task-forwarder")
-            .expect("runtime present")
-            .state
-            .scripts[0]
-            .buffered_terminal_chunks
-            .iter()
-            .map(|chunk| chunk.data.as_str())
-            .collect::<String>()
-    }
-
-    #[test]
-    fn sync_group_state_updates_order_and_retains_live_removed_scripts() {
-        let mut state = DevServerGroupState {
-            repo_path: "/repo".to_string(),
-            task_id: "task-1".to_string(),
-            worktree_path: Some("/repo/worktree".to_string()),
-            scripts: vec![
-                DevServerScriptState {
-                    script_id: "frontend".to_string(),
-                    name: "Frontend old".to_string(),
-                    command: "bun run dev:old".to_string(),
-                    status: DevServerScriptStatus::Stopped,
-                    pid: None,
-                    started_at: None,
-                    exit_code: None,
-                    last_error: None,
-                    buffered_terminal_chunks: vec![DevServerTerminalChunk {
-                        script_id: "frontend".to_string(),
-                        sequence: 0,
-                        data: "kept\r\n".to_string(),
-                        timestamp: "2026-03-19T10:00:00Z".to_string(),
-                    }],
-                    next_terminal_sequence: 1,
-                },
-                DevServerScriptState {
-                    script_id: "orphan".to_string(),
-                    name: "Old orphan".to_string(),
-                    command: "sleep 30".to_string(),
-                    status: DevServerScriptStatus::Running,
-                    pid: Some(4242),
-                    started_at: Some("2026-03-19T10:01:00Z".to_string()),
-                    exit_code: None,
-                    last_error: None,
-                    buffered_terminal_chunks: Vec::new(),
-                    next_terminal_sequence: 0,
-                },
-            ],
-            updated_at: "2026-03-19T10:00:00Z".to_string(),
-        };
-
-        sync_group_state(
-            &mut state,
-            "/repo",
-            "task-1",
-            Some("/repo/worktree-next".to_string()),
-            &repo_config(vec![
-                RepoDevServerScript {
-                    id: "backend".to_string(),
-                    name: "Backend".to_string(),
-                    command: "bun run api".to_string(),
-                },
-                RepoDevServerScript {
-                    id: "frontend".to_string(),
-                    name: "Frontend".to_string(),
-                    command: "bun run web".to_string(),
-                },
-            ]),
-        );
-
-        assert_eq!(state.worktree_path.as_deref(), Some("/repo/worktree-next"));
-        assert_eq!(state.scripts.len(), 3);
-        assert_eq!(state.scripts[0].script_id, "backend");
-        assert_eq!(state.scripts[0].name, "Backend");
-        assert_eq!(state.scripts[1].script_id, "frontend");
-        assert_eq!(state.scripts[1].name, "Frontend");
-        assert_eq!(state.scripts[1].command, "bun run web");
-        assert_eq!(state.scripts[1].buffered_terminal_chunks.len(), 1);
-        assert_eq!(state.scripts[2].script_id, "orphan");
-        assert_eq!(state.scripts[2].pid, Some(4242));
-    }
-
-    #[test]
-    fn emit_terminal_chunk_trims_buffer_and_emits_events() {
-        let events = Arc::new(Mutex::new(Vec::new()));
-        let emitter_events = events.clone();
-        let emitter: DevServerEmitter = Arc::new(move |event| {
-            emitter_events
-                .lock()
-                .expect("event lock poisoned")
-                .push(event);
-        });
-        let groups = Arc::new(Mutex::new(HashMap::from([(
-            "repo::task-1".to_string(),
-            DevServerGroupRuntime {
-                state: DevServerGroupState {
-                    repo_path: "repo".to_string(),
-                    task_id: "task-1".to_string(),
-                    worktree_path: Some("/tmp/worktree".to_string()),
-                    scripts: vec![DevServerScriptState {
-                        script_id: "server-1".to_string(),
-                        name: "Server".to_string(),
-                        command: "bun run dev".to_string(),
-                        status: DevServerScriptStatus::Running,
-                        pid: Some(99),
-                        started_at: Some("2026-03-19T10:00:00Z".to_string()),
-                        exit_code: None,
-                        last_error: None,
-                        buffered_terminal_chunks: Vec::new(),
-                        next_terminal_sequence: 0,
-                    }],
-                    updated_at: "2026-03-19T10:00:00Z".to_string(),
-                },
-                emitter: Some(emitter),
-            },
-        )])));
-
-        for index in 0..(DEV_SERVER_TERMINAL_BUFFER_CHUNK_LIMIT + 5) {
-            emit_terminal_chunk(
-                &groups,
-                "repo::task-1",
-                "repo",
-                "task-1",
-                "server-1",
-                format!("line-{index}"),
-            );
-        }
-
-        let groups = groups.lock().expect("group lock poisoned");
-        let runtime = groups.get("repo::task-1").expect("runtime present");
-        let logs = &runtime.state.scripts[0].buffered_terminal_chunks;
-        assert_eq!(logs.len(), DEV_SERVER_TERMINAL_BUFFER_CHUNK_LIMIT);
-        assert_eq!(
-            logs.first().map(|chunk| chunk.data.as_str()),
-            Some("line-5")
-        );
-        assert_eq!(
-            logs.last().map(|chunk| chunk.data.as_str()),
-            Some("line-2004")
-        );
-        drop(groups);
-
-        let emitted = events.lock().expect("event lock poisoned");
-        assert_eq!(emitted.len(), DEV_SERVER_TERMINAL_BUFFER_CHUNK_LIMIT + 5);
-        assert!(matches!(
-            emitted.last(),
-            Some(DevServerEvent::TerminalChunk { terminal_chunk, .. }) if terminal_chunk.data == "line-2004"
-        ));
-    }
-
-    #[test]
-    fn emit_terminal_chunk_trims_buffer_by_byte_budget() {
-        let groups = Arc::new(Mutex::new(HashMap::from([(
-            "repo::task-1".to_string(),
-            DevServerGroupRuntime {
-                state: DevServerGroupState {
-                    repo_path: "repo".to_string(),
-                    task_id: "task-1".to_string(),
-                    worktree_path: Some("/tmp/worktree".to_string()),
-                    scripts: vec![DevServerScriptState {
-                        script_id: "server-1".to_string(),
-                        name: "Server".to_string(),
-                        command: "bun run dev".to_string(),
-                        status: DevServerScriptStatus::Running,
-                        pid: Some(99),
-                        started_at: Some("2026-03-19T10:00:00Z".to_string()),
-                        exit_code: None,
-                        last_error: None,
-                        buffered_terminal_chunks: Vec::new(),
-                        next_terminal_sequence: 0,
-                    }],
-                    updated_at: "2026-03-19T10:00:00Z".to_string(),
-                },
-                emitter: None,
-            },
-        )])));
-
-        emit_terminal_chunk(
-            &groups,
-            "repo::task-1",
-            "repo",
-            "task-1",
-            "server-1",
-            "a".repeat(DEV_SERVER_TERMINAL_BUFFER_BYTE_LIMIT - 32),
-        );
-        emit_terminal_chunk(
-            &groups,
-            "repo::task-1",
-            "repo",
-            "task-1",
-            "server-1",
-            "b".repeat(128),
-        );
-
-        let groups = groups.lock().expect("group lock poisoned");
-        let runtime = groups.get("repo::task-1").expect("runtime present");
-        let buffered = &runtime.state.scripts[0].buffered_terminal_chunks;
-        assert_eq!(buffered.len(), 1);
-        assert!(buffered[0].data.starts_with('b'));
-    }
-
-    #[test]
-    fn spawn_terminal_forwarder_retries_interrupted_reads() {
-        let groups = build_runtime_groups_for_forwarder_test();
-
-        spawn_terminal_forwarder(
-            groups.clone(),
-            "repo::task-forwarder".to_string(),
-            "repo".to_string(),
-            "task-forwarder".to_string(),
-            "server-1".to_string(),
-            ScriptedReader::new(vec![
-                ScriptedReadStep::Error(io::ErrorKind::Interrupted),
-                ScriptedReadStep::Chunk(b"ready\r\n".to_vec()),
-                ScriptedReadStep::Eof,
-            ]),
-        );
-
-        let mut captured = String::new();
-        for _ in 0..20 {
-            thread::sleep(Duration::from_millis(25));
-            captured = read_forwarder_output(&groups);
-            if captured.contains("ready\r\n") {
-                break;
-            }
-        }
-
-        assert!(captured.contains("ready\r\n"));
-        assert!(!captured.contains("terminal stream failed"));
-    }
-
-    #[test]
-    fn spawn_terminal_forwarder_skips_invalid_utf8_and_continues_streaming() {
-        let groups = build_runtime_groups_for_forwarder_test();
-
-        spawn_terminal_forwarder(
-            groups.clone(),
-            "repo::task-forwarder".to_string(),
-            "repo".to_string(),
-            "task-forwarder".to_string(),
-            "server-1".to_string(),
-            ScriptedReader::new(vec![
-                ScriptedReadStep::Chunk(vec![0xff]),
-                ScriptedReadStep::Chunk(b"ready\r\n".to_vec()),
-                ScriptedReadStep::Eof,
-            ]),
-        );
-
-        let mut captured = String::new();
-        for _ in 0..20 {
-            thread::sleep(Duration::from_millis(25));
-            captured = read_forwarder_output(&groups);
-            if captured.contains("ready\r\n") {
-                break;
-            }
-        }
-
-        assert!(captured.contains(
-            "Dev server terminal output contained invalid UTF-8 bytes and could not be fully rendered."
-        ));
-        assert!(captured.contains("ready\r\n"));
-    }
-
-    #[test]
-    fn start_dev_server_script_rejects_blank_commands() {
-        let (service, _task_state, _git_state) = build_service_with_state(Vec::new());
-        let repo_path = "/repo";
-        let task_id = "task-parse";
-        let group_key = dev_server_group_key(repo_path, task_id);
-        let script = RepoDevServerScript {
-            id: "frontend".to_string(),
-            name: "Frontend".to_string(),
-            command: "   ".to_string(),
-        };
-
-        service
-            .dev_server_groups
-            .lock()
-            .expect("group lock poisoned")
-            .insert(
-                group_key.clone(),
-                DevServerGroupRuntime {
-                    state: build_group_state(
-                        repo_path,
-                        task_id,
-                        Some("/tmp/worktree".to_string()),
-                        &repo_config(vec![script.clone()]),
-                    ),
-                    emitter: None,
-                },
-            );
-
-        let error = service
-            .start_dev_server_script(
-                group_key.as_str(),
-                repo_path,
-                task_id,
-                "/tmp/worktree",
-                script,
-            )
-            .expect_err("blank command should fail");
-
-        assert!(error
-            .to_string()
-            .contains("Dev server command is empty for Frontend"));
-
-        let groups = service
-            .dev_server_groups
-            .lock()
-            .expect("group lock poisoned");
-        let runtime = groups.get(&group_key).expect("runtime present");
-        let script = &runtime.state.scripts[0];
-        assert_eq!(script.status, DevServerScriptStatus::Failed);
-        assert_eq!(script.pid, None);
-        assert_eq!(script.started_at, None);
-        assert_eq!(script.exit_code, None);
-        assert!(matches!(
-            script.last_error.as_deref(),
-            Some(message) if message.contains("Dev server command is empty for Frontend")
-        ));
-        assert!(script.buffered_terminal_chunks.iter().any(|chunk| chunk
-            .data
-            .contains("Dev server command is empty for Frontend")));
-    }
-
-    #[test]
-    fn dev_server_start_requires_builder_worktree_before_hook_trust() {
-        let root = unique_temp_path("dev-server-start-missing-worktree");
-        let repo = root.join("repo");
-        init_git_repo(&repo).expect("test repo should initialize");
-
-        let config_store = AppConfigStore::from_path(root.join("config.json"));
-        let (service, _task_state, _git_state) = build_service_with_store(
-            vec![make_task("task-1", "task", TaskStatus::InProgress)],
-            vec![],
-            GitCurrentBranch {
-                name: Some("main".to_string()),
-                detached: false,
-                revision: None,
-            },
-            config_store,
-        );
-        let repo_path = repo.to_string_lossy().to_string();
-        service
-            .workspace_add(repo_path.as_str())
-            .expect("workspace should add");
-        service
-            .workspace_update_repo_config(
-                repo_path.as_str(),
-                RepoConfig {
-                    dev_servers: vec![RepoDevServerScript {
-                        id: "frontend".to_string(),
-                        name: "Frontend".to_string(),
-                        command: "bun run dev".to_string(),
-                    }],
-                    trusted_hooks: false,
-                    trusted_hooks_fingerprint: None,
-                    ..Default::default()
-                },
-            )
-            .expect("repo config should persist");
-
-        let emitter: DevServerEmitter = Arc::new(|_| {});
-        let error = service
-            .dev_server_start(repo_path.as_str(), "task-1", emitter)
-            .expect_err("missing builder worktree should fail before hook trust");
-        assert_eq!(
-            error.to_string(),
-            "Builder continuation cannot start until a builder worktree exists for task task-1. Start Builder first."
-        );
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn start_dev_server_script_reports_immediate_shell_failures() {
-        let (service, _task_state, _git_state) = build_service_with_state(Vec::new());
-        let repo_path = "/repo";
-        let task_id = "task-spawn";
-        let worktree_path = unique_temp_path("dev-server-worktree");
-        fs::create_dir_all(&worktree_path).expect("create worktree path");
-        let worktree_path = worktree_path.to_string_lossy().to_string();
-        let group_key = dev_server_group_key(repo_path, task_id);
-        let script = RepoDevServerScript {
-            id: "backend".to_string(),
-            name: "Backend".to_string(),
-            command: "__odt_missing_executable__ --port 3000".to_string(),
-        };
-
-        service
-            .dev_server_groups
-            .lock()
-            .expect("group lock poisoned")
-            .insert(
-                group_key.clone(),
-                DevServerGroupRuntime {
-                    state: build_group_state(
-                        repo_path,
-                        task_id,
-                        Some(worktree_path.clone()),
-                        &repo_config(vec![script.clone()]),
-                    ),
-                    emitter: None,
-                },
-            );
-
-        let error = service
-            .start_dev_server_script(
-                group_key.as_str(),
-                repo_path,
-                task_id,
-                worktree_path.as_str(),
-                script,
-            )
-            .expect_err("missing command should fail during startup");
-
-        assert!(error
-            .to_string()
-            .contains("Dev server exited with code 127"));
-
-        let groups = service
-            .dev_server_groups
-            .lock()
-            .expect("group lock poisoned");
-        let runtime = groups.get(&group_key).expect("runtime present");
-        let script = &runtime.state.scripts[0];
-        assert_eq!(script.status, DevServerScriptStatus::Failed);
-        assert_eq!(script.pid, None);
-        assert_eq!(script.started_at, None);
-        assert_eq!(script.exit_code, Some(127));
-        assert!(matches!(
-            script.last_error.as_deref(),
-            Some(message) if message.contains("Dev server exited with code")
-        ));
-        assert!(script
-            .buffered_terminal_chunks
-            .iter()
-            .any(|chunk| chunk.data.contains("Dev server exited with code")));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn dev_server_start_keeps_successful_scripts_running_when_another_script_fails() {
-        let root = unique_temp_path("dev-server-partial-start");
-        let repo = root.join("repo");
-        init_git_repo(&repo).expect("test repo should initialize");
-
-        let config_store = AppConfigStore::from_path(root.join("config.json"));
-        let (service, _task_state, _git_state) = build_service_with_store(
-            vec![make_task("task-1", "task", TaskStatus::InProgress)],
-            vec![],
-            GitCurrentBranch {
-                name: Some("main".to_string()),
-                detached: false,
-                revision: None,
-            },
-            config_store,
-        );
-        let repo_path = repo.to_string_lossy().to_string();
-        let worktree_path = repo.join(".openducktor-worktree");
-        fs::create_dir_all(&worktree_path).expect("create worktree path");
-        let worktree_path = worktree_path.to_string_lossy().to_string();
-
-        service
-            .workspace_add(repo_path.as_str())
-            .expect("workspace should add");
-        service
-            .workspace_update_repo_config(
-                repo_path.as_str(),
-                RepoConfig {
-                    trusted_hooks: true,
-                    dev_servers: vec![
-                        RepoDevServerScript {
-                            id: "frontend".to_string(),
-                            name: "Frontend".to_string(),
-                            command: "sleep 20".to_string(),
-                        },
-                        RepoDevServerScript {
-                            id: "backend".to_string(),
-                            name: "Backend".to_string(),
-                            command: "__odt_missing_executable__ --port 3000".to_string(),
-                        },
-                    ],
-                    ..Default::default()
-                },
-            )
-            .expect("repo config should persist");
-        let challenge = service
-            .workspace_prepare_trusted_hooks_challenge(repo_path.as_str())
-            .expect("trust challenge should prepare");
-        service
-            .workspace_set_trusted_hooks(
-                repo_path.as_str(),
-                true,
-                Some(challenge.nonce.as_str()),
-                Some(challenge.fingerprint.as_str()),
-                &AllowHookTrustConfirmationPort,
-            )
-            .expect("hook trust should persist");
-
-        let mut session = crate::app_service::test_support::make_session("task-1", "build-session");
-        session.role = "build".to_string();
-        session.working_directory = worktree_path;
-        service
-            .agent_session_upsert(repo_path.as_str(), "task-1", session)
-            .expect("builder session should persist");
-
-        let emitter: DevServerEmitter = Arc::new(|_| {});
-        let state = service
-            .dev_server_start(repo_path.as_str(), "task-1", emitter)
-            .expect("partial dev server start should keep successful scripts running");
-
-        let frontend = state
-            .scripts
-            .iter()
-            .find(|script| script.script_id == "frontend")
-            .expect("frontend script present");
-        assert_eq!(frontend.status, DevServerScriptStatus::Running);
-        assert!(frontend.pid.is_some());
-
-        let backend = state
-            .scripts
-            .iter()
-            .find(|script| script.script_id == "backend")
-            .expect("backend script present");
-        assert_eq!(backend.status, DevServerScriptStatus::Failed);
-        assert_eq!(backend.exit_code, Some(127));
-        assert!(matches!(
-            backend.last_error.as_deref(),
-            Some(message) if message.contains("Dev server exited with code 127")
-        ));
-
-        let stopped = service
-            .dev_server_stop(repo_path.as_str(), "task-1")
-            .expect("running scripts should stop cleanly");
-        let stopped_frontend = stopped
-            .scripts
-            .iter()
-            .find(|script| script.script_id == "frontend")
-            .expect("frontend script present after stop");
-        assert_eq!(stopped_frontend.status, DevServerScriptStatus::Stopped);
-        assert_eq!(stopped_frontend.pid, None);
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn start_dev_server_script_uses_augmented_subprocess_path() {
-        let _env_lock = lock_env();
-        let (service, _task_state, _git_state) = build_service_with_state(Vec::new());
-        let repo_path = "/repo";
-        let task_id = "task-path";
-        let sandbox = unique_temp_path("dev-server-path");
-        let home = sandbox.join("home");
-        let login_bin = sandbox.join("login-bin");
-        fs::create_dir_all(&login_bin).expect("create login shell path");
-        let pnpm_path = login_bin.join("pnpm");
-        let node_path = login_bin.join("node");
-        let shell_path = sandbox.join("fake-shell");
-        let marker_path = sandbox.join("pnpm-marker");
-        write_executable_script(
-            pnpm_path.as_path(),
-            "#!/usr/bin/env node\nconsole.log('pnpm shim');\n",
-        )
-        .expect("write fake pnpm");
-        write_executable_script(
-            node_path.as_path(),
-            format!(
-                r#"#!/bin/sh
-touch "{}"
-sleep 5
-"#,
-                marker_path.display()
-            )
-            .as_str(),
-        )
-        .expect("write fake node");
-        write_executable_script(
-            shell_path.as_path(),
-            format!(
-                "#!/bin/sh\nprintf 'shell startup noise\\n'\nprintf '__OPENDUCKTOR_ENV_START__\\0PATH={}\\0'\n",
-                login_bin.display()
-            )
-            .as_str(),
-        )
-        .expect("write fake shell");
-
-        let worktree_path = sandbox.join("worktree");
-        fs::create_dir_all(&worktree_path).expect("create worktree path");
-        let worktree_path = worktree_path.to_string_lossy().to_string();
-        let group_key = dev_server_group_key(repo_path, task_id);
-        let script = RepoDevServerScript {
-            id: "frontend".to_string(),
-            name: "Frontend".to_string(),
-            command: "pnpm".to_string(),
-        };
-
-        service
-            .dev_server_groups
-            .lock()
-            .expect("group lock poisoned")
-            .insert(
-                group_key.clone(),
-                DevServerGroupRuntime {
-                    state: build_group_state(
-                        repo_path,
-                        task_id,
-                        Some(worktree_path.clone()),
-                        &repo_config(vec![script.clone()]),
-                    ),
-                    emitter: None,
-                },
-            );
-
-        let _home_guard = set_env_var("HOME", home.to_string_lossy().as_ref());
-        let _shell_guard = set_env_var("SHELL", shell_path.to_string_lossy().as_ref());
-        let _user_guard = set_env_var("USER", "odt-test");
-        let _logname_guard = set_env_var("LOGNAME", "odt-test");
-        let _path_guard = set_env_var("PATH", "/usr/bin:/bin");
-
-        let pid = service
-            .start_dev_server_script(
-                group_key.as_str(),
-                repo_path,
-                task_id,
-                worktree_path.as_str(),
-                script,
-            )
-            .expect("fake pnpm and node from augmented PATH should start");
-
-        assert!(
-            wait_for_path_exists(marker_path.as_path(), Duration::from_secs(5)),
-            "expected fake pnpm script to resolve node from augmented PATH"
-        );
-        stop_process_group(pid, DEV_SERVER_STOP_TIMEOUT).expect("stop dev server");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn start_dev_server_script_streams_logs_before_process_exit_and_clears_old_logs() {
-        let (service, _task_state, _git_state) = build_service_with_state(Vec::new());
-        let repo_path = "/repo";
-        let task_id = "task-stream";
-        let worktree_path = unique_temp_path("dev-server-stream-worktree");
-        fs::create_dir_all(&worktree_path).expect("create worktree path");
-        let worktree_path = worktree_path.to_string_lossy().to_string();
-        let group_key = dev_server_group_key(repo_path, task_id);
-        let script = RepoDevServerScript {
-            id: "frontend".to_string(),
-            name: "Frontend".to_string(),
-            command: "printf 'db generated\\n' && python3 -c \"import time; print('ready'); time.sleep(5)\""
-                .to_string(),
-        };
-
-        let mut state = build_group_state(
-            repo_path,
-            task_id,
-            Some(worktree_path.clone()),
-            &repo_config(vec![script.clone()]),
-        );
-        state.scripts[0]
-            .buffered_terminal_chunks
-            .push(DevServerTerminalChunk {
-                script_id: "frontend".to_string(),
-                sequence: 0,
-                data: "stale log\r\n".to_string(),
-                timestamp: "2026-03-19T10:00:00Z".to_string(),
-            });
-        state.scripts[0].next_terminal_sequence = 1;
-
-        service
-            .dev_server_groups
-            .lock()
-            .expect("group lock poisoned")
-            .insert(
-                group_key.clone(),
-                DevServerGroupRuntime {
-                    state,
-                    emitter: None,
-                },
-            );
-
-        service
-            .start_dev_server_script(
-                group_key.as_str(),
-                repo_path,
-                task_id,
-                worktree_path.as_str(),
-                script,
-            )
-            .expect("dev server should start");
-
-        let mut saw_ready_log = false;
-        let mut saw_setup_log = false;
-        let mut pid = None;
-        for _ in 0..30 {
-            thread::sleep(Duration::from_millis(100));
-            let groups = service
-                .dev_server_groups
-                .lock()
-                .expect("group lock poisoned");
-            let runtime = groups.get(&group_key).expect("runtime present");
-            let script = &runtime.state.scripts[0];
-            pid = script.pid;
-            assert!(script
-                .buffered_terminal_chunks
-                .iter()
-                .all(|chunk| chunk.data != "stale log\r\n"));
-            saw_setup_log = script
-                .buffered_terminal_chunks
-                .iter()
-                .any(|chunk| chunk.data.contains("db generated"));
-            saw_ready_log = script
-                .buffered_terminal_chunks
-                .iter()
-                .any(|chunk| chunk.data.contains("ready"));
-            if saw_setup_log && saw_ready_log {
-                break;
-            }
-        }
-
-        assert!(
-            saw_setup_log,
-            "expected chained shell setup log before process exit"
-        );
-        assert!(saw_ready_log, "expected log line before process exit");
-
-        let pid = pid.expect("dev server pid missing");
-        stop_process_group(pid, DEV_SERVER_STOP_TIMEOUT).expect("stop streamed dev server");
-        for _ in 0..30 {
-            thread::sleep(Duration::from_millis(50));
-            let groups = service
-                .dev_server_groups
-                .lock()
-                .expect("group lock poisoned");
-            let runtime = groups.get(&group_key).expect("runtime present");
-            if runtime.state.scripts[0].pid.is_none() {
-                break;
-            }
-        }
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn start_dev_server_script_preserves_terminal_sequences_and_merges_stderr() {
-        let (service, _task_state, _git_state) = build_service_with_state(Vec::new());
-        let repo_path = "/repo";
-        let task_id = "task-terminal";
-        let worktree_path = unique_temp_path("dev-server-terminal-worktree");
-        fs::create_dir_all(&worktree_path).expect("create worktree path");
-        let worktree_path = worktree_path.to_string_lossy().to_string();
-        let group_key = dev_server_group_key(repo_path, task_id);
-        let script = RepoDevServerScript {
-            id: "frontend".to_string(),
-            name: "Frontend".to_string(),
-            command:
-                "printf '\\033[32mready\\033[0m\\r' && printf 'stderr info\\r\\n' >&2 && sleep 5"
-                    .to_string(),
-        };
-
-        service
-            .dev_server_groups
-            .lock()
-            .expect("group lock poisoned")
-            .insert(
-                group_key.clone(),
-                DevServerGroupRuntime {
-                    state: build_group_state(
-                        repo_path,
-                        task_id,
-                        Some(worktree_path.clone()),
-                        &repo_config(vec![script.clone()]),
-                    ),
-                    emitter: None,
-                },
-            );
-
-        service
-            .start_dev_server_script(
-                group_key.as_str(),
-                repo_path,
-                task_id,
-                worktree_path.as_str(),
-                script,
-            )
-            .expect("dev server should start");
-
-        let mut pid = None;
-        let mut captured = String::new();
-        for _ in 0..30 {
-            thread::sleep(Duration::from_millis(100));
-            let groups = service
-                .dev_server_groups
-                .lock()
-                .expect("group lock poisoned");
-            let runtime = groups.get(&group_key).expect("runtime present");
-            let script = &runtime.state.scripts[0];
-            pid = script.pid;
-            captured = script
-                .buffered_terminal_chunks
-                .iter()
-                .map(|chunk| chunk.data.as_str())
-                .collect::<String>();
-            if captured.contains("stderr info") {
-                break;
-            }
-        }
-
-        assert!(captured.contains("\u{1b}[32mready\u{1b}[0m\r"));
-        assert!(captured.contains("stderr info"));
-
-        let pid = pid.expect("dev server pid missing");
-        stop_process_group(pid, DEV_SERVER_STOP_TIMEOUT).expect("stop terminal dev server");
-    }
-
-    #[test]
-    fn mark_dev_servers_stopping_clears_terminal_replay_and_resets_failed_scripts() {
-        let (service, _task_state, _git_state) = build_service_with_state(Vec::new());
-        let group_key = "repo::task-stop".to_string();
-        service
-            .dev_server_groups
-            .lock()
-            .expect("group lock poisoned")
-            .insert(
-                group_key.clone(),
-                DevServerGroupRuntime {
-                    state: DevServerGroupState {
-                        repo_path: "repo".to_string(),
-                        task_id: "task-stop".to_string(),
-                        worktree_path: Some("/tmp/worktree".to_string()),
-                        scripts: vec![
-                            DevServerScriptState {
-                                script_id: "frontend".to_string(),
-                                name: "Frontend".to_string(),
-                                command: "bun run dev".to_string(),
-                                status: DevServerScriptStatus::Running,
-                                pid: Some(4242),
-                                started_at: Some("2026-03-19T10:00:00Z".to_string()),
-                                exit_code: None,
-                                last_error: None,
-                                buffered_terminal_chunks: vec![DevServerTerminalChunk {
-                                    script_id: "frontend".to_string(),
-                                    sequence: 0,
-                                    data: "ready\r\n".to_string(),
-                                    timestamp: "2026-03-19T10:00:00Z".to_string(),
-                                }],
-                                next_terminal_sequence: 1,
-                            },
-                            DevServerScriptState {
-                                script_id: "backend".to_string(),
-                                name: "Backend".to_string(),
-                                command: "bun run api".to_string(),
-                                status: DevServerScriptStatus::Failed,
-                                pid: None,
-                                started_at: None,
-                                exit_code: Some(1),
-                                last_error: Some("boom".to_string()),
-                                buffered_terminal_chunks: vec![DevServerTerminalChunk {
-                                    script_id: "backend".to_string(),
-                                    sequence: 0,
-                                    data: "boom\r\n".to_string(),
-                                    timestamp: "2026-03-19T10:00:01Z".to_string(),
-                                }],
-                                next_terminal_sequence: 1,
-                            },
-                        ],
-                        updated_at: "2026-03-19T10:00:00Z".to_string(),
-                    },
-                    emitter: None,
-                },
-            );
-
-        let targets = service
-            .mark_dev_servers_stopping(group_key.as_str())
-            .expect("stop mark should succeed");
-
-        assert_eq!(targets, vec![("frontend".to_string(), 4242)]);
-
-        let groups = service
-            .dev_server_groups
-            .lock()
-            .expect("group lock poisoned");
-        let runtime = groups.get(&group_key).expect("runtime present");
-        assert_eq!(
-            runtime.state.scripts[0].status,
-            DevServerScriptStatus::Stopping
-        );
-        assert!(runtime.state.scripts[0].buffered_terminal_chunks.is_empty());
-        assert_eq!(
-            runtime.state.scripts[1].status,
-            DevServerScriptStatus::Stopped
-        );
-        assert_eq!(runtime.state.scripts[1].last_error, None);
-        assert_eq!(runtime.state.scripts[1].exit_code, None);
-        assert!(runtime.state.scripts[1].buffered_terminal_chunks.is_empty());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn dev_server_parent_death_watcher_terminates_orphaned_process_group() {
-        let mut child = spawn_sleep_process_group(20);
-        let pid = child.id();
-
-        spawn_dev_server_parent_death_watcher(999_999, pid)
-            .expect("watcher should start for dev server process");
-
-        assert!(
-            wait_for_process_exit(pid as i32, Duration::from_secs(3)),
-            "dev server process group should exit when parent is already gone"
-        );
-        let _ = child.wait().expect("failed waiting dev server child");
-    }
-}
